backend/requirements.txt
llama-index
faiss-cpu>=1.8.0
requests
beautifulsoup4
pypdf
python-dotenv
tqdm
numpy
sentence-transformers
llama-index-embeddings-huggingface
fastmcp
yfinance 
pandas
langchain
langchain-groq
langchain==0.2.12 
langchain-core==0.2.30 
langchain-community==0.2.11 
fastapi
langsmith